renamed dataGeojson action creator and reducer to coverage
